multi: remove old logger & restructure
- Remove the old logger implementation.
- Move various helper methods to util.go
- Rename slog.go to log.go
- remove travis config file
